code/src/requirements.txt
PyYAML>=6.0
pycountry
requests
pandas
streamlit
pandas
newspaper3k
nltk
pyvis
fastapi
python-dotenv
gunicorn
apache-airflow
lxml_html_clean
google-generativeai>=0.3.0
SPARQLWrapper
uvicorn
httpx
python-dotenv
neo4j